moved more output code to a formatter module
